+++ lib/SMLParser/sml.h
@@ -97,10 +97,9 @@ bool smlOBISCheck(const unsigned char *obis);
 void smlOBISManufacturer(unsigned char *str, int maxSize);
 void smlOBISByUnit(long long int &wh, signed char &scaler, sml_units_t unit);
 
-// Be aware that double on Arduino UNO is just 32 bit
-void smlOBISWh(double &wh);
-void smlOBISW(double &w);
-void smlOBISVolt(double &v);
-void smlOBISAmpere(double &a);
+void smlOBISWh(float &wh);
+void smlOBISW(float &w);
+void smlOBISVolt(float &v);
+void smlOBISAmpere(float &a);
 
 #endif
